+++ cerulean/local_file_system.py
@@ -1,8 +1,9 @@
 import errno
 import os
 import pathlib
-from typing import cast, Any, Generator, Iterable
+from typing import cast, Any, Generator, Iterable, Optional
 
+from cerulean.file_system import FileSystem
 from cerulean.file_system_impl import FileSystemImpl
 from cerulean.path import AbstractPath, EntryType, Path, Permission
 
@@ -27,6 +28,11 @@ class LocalFileSystem(FileSystemImpl):
     be good to do so anyway, to avoid leaks if you ever replace it with \
     a different :class:`FileSystem` that does.
     """
+    def __eq__(self, other: Any) -> bool:
+        if not isinstance(other, FileSystem):
+            return NotImplemented
+        return isinstance(other, LocalFileSystem)
+
     def root(self) -> Path:
         return Path(self, pathlib.Path('/'))
 
@@ -52,11 +58,14 @@ class LocalFileSystem(FileSystemImpl):
 
     def _mkdir(self,
               path: AbstractPath,
-              mode: int = 0o777,
+              mode: Optional[int] = None,
               parents: bool = False,
               exists_ok: bool = False) -> None:
         lpath = cast(pathlib.Path, path)
-        lpath.mkdir(mode, parents, exists_ok)
+        if mode is not None:
+            lpath.mkdir(mode, parents, exists_ok)
+        else:
+            lpath.mkdir(parents=parents, exist_ok=exists_ok)
 
     def _iterdir(self, path: AbstractPath) -> Generator[AbstractPath, None, None]:
         lpath = cast(pathlib.Path, path)