+++ cerulean/scheduler.py
@@ -67,32 +67,46 @@ class Scheduler(ABC):
         """
         pass
 
-    def wait(self, job_id: str, time_out: float = -1.0) -> Optional[int]:
+    def wait(self, job_id: str, time_out: float = -1.0, interval: float = None
+             ) -> Optional[int]:
         """Wait until the job is done.
 
         Will wait approximately time_out seconds for the job to finish.
         Returns the exit code if the job finished, otherwise None.
-        Resolution is 0.1s, or time_out / 20, whichever is larger.
+
+        The state of the job will be checked every `interval` seconds.
+        If `interval` is None, or not specified, then the interval will
+        be 1s, or time_out / 50, whichever is larger. If time_out is
+        not given, the interval will start at 1s, then increase
+        gradually to about 30s.
 
         Args:
             job_id: The job to wait for.
             time_out: Time to wait in seconds. If negative, wait \
                     forever.
+            interval: Time to wait between checks. See above.
 
         Returns:
             The exit code of the job.
         """
         if time_out < 0.0:
             time_end = None
+            fixed_interval = False
         else:
             time_end = perf_counter() + time_out
+            fixed_interval = True
 
-        sleep_time = max(0.1, time_out / 20.0)
+        if interval is None:
+            # this is correct for negative time-outs as well
+            interval = max(1.0, time_out / 50.0)
 
         status = self.get_status(job_id)
         while status != JobStatus.DONE:
-            sleep(sleep_time)
+            sleep(interval)
             status = self.get_status(job_id)
+
+            if not fixed_interval and interval < 30:
+                interval += 2
 
             if time_end is not None and time_end < perf_counter():
                 break